requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
limits==3.6.0
